feat: use requestData for upcoming and latest launch
